Moving LoginManager to run.py
